adapt high-level wrappers to MEOS 1.3 API
MEOS 1.3 makes three sweeping shape changes that the regenerated
bindings expose:

1. Every predicate that used to return `bool` now returns `int`
   (0/1). 86 call sites under MEOS.NET/Types/ wrapped with
   `!= 0` so each caller's domain bool semantics stays explicit
   instead of hidden under a generated wrapper.

2. Every `bool` parameter (span_make's lower_inc/upper_inc,
   asMfJson's bounding-box flag, temporal_duration's ignoreGaps,
   ...) flipped to `int`. 29 call sites cast `b ? 1 : 0` at the
   boundary.

3. Several helpers renamed (`tfloat_round` ->
   `temporal_round`, `tfloat_derivative` -> `temporal_derivative`,
   `tfloat_ever_eq` -> `ever_eq_tfloat_float` and family,
   `span_width` -> `numspan_width`, `spanset_width` ->
   `numspanset_width`, `distance_tnumber_tnumber` ->
   `tdistance_tnumber_tnumber`, `distance_tfloat_float` ->
   `tdistance_tfloat_float`, `tstzspan_to_tbox` -> `span_to_tbox`,
   `tstzset_to_tbox` -> `set_to_tbox`, `tstzspanset_to_tbox` ->
   `spanset_to_tbox`).  Call sites updated directly; no compat
   aliases in the codegen layer.

Also:

- `meos_initialize` is now zero-argument. `MEOSLifecycle.Initialize`
  splits into three calls (`meos_initialize`,
  `meos_initialize_timezone`, `meos_initialize_error_handler`) and
  marshals the C# error-handler delegate to a function pointer via
  `Marshal.GetFunctionPointerForDelegate`, keeping a static
  reference alive to prevent GC collection while MEOS holds the
  pointer.
- `temporal_from_mfjson` now requires a `meosType` enum; the
  generic `Temporal.FromMFJson(string)` factory cannot pick a
  subtype from JSON content alone, so it was removed.  Subtype-
  specific factories live on subclasses (TemporalBoolean.FromMFJson
  via `tbool_from_mfjson`, TemporalFloat.FromMFJson via
  `tfloat_from_mfjson`).
- `*_from_wkb` size argument is `int` (was `ulong`).  Four call
  sites drop the `(ulong)` cast.
- `tfloatseq_from_base_tstzspan` and the sequence-set form take an
  `interpType` enum; `MEOS.NET.Enums.InterpolationType` casts
  explicitly with `(int)`.

Solution builds 0-error / 0-warning across MEOS.NET,
MEOS.NET.Builder, MEOS.NET.NpgSql, ExampleApp, MEOS.NET.Tests.

+++ MEOS.NET/Types/Temporal/Number/Float/TemporalFloatSequence.cs
@@ -19,7 +19,7 @@ namespace MEOS.NET.Types.Temporal.Number.Float
 
         public static TemporalFloatSequence FromTimestampSpan(TimestampTzSpan tsTzSpan, double value, InterpolationType interpolation = InterpolationType.None)
         {
-            var res = MEOSExposedFunctions.tfloatseq_from_base_tstzspan(value, tsTzSpan._ptr, interpolation);
+            var res = MEOSExposedFunctions.tfloatseq_from_base_tstzspan(value, tsTzSpan._ptr, (int)interpolation);
             return new TemporalFloatSequence(res);
         }
     }
